feat: Add C library for GByteArray mapping
Most of `GArray` (and by extension `GByteArray`) is not introspectable,
so the mapping has to be created manually.
This requires switching the Rockspec to Make, so I also ditched Just.
A name change to `lgi_async_extra` is also required, since the C library
cannot be referenced with a name containing dashes.